requirements.txt
livekit
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-turn-detector
vosk
python-dotenv
pyyaml
